Report
Andrew Zannetos & Cassie Loe
EC463
9/20/18
Hardware Project Report
Operating System
A micro SD card was bought, OS installed onto the SD, however, we were unable to communicate with the Raspberry Pi due to the problems outlined below.

 
“Connecting” the Raspberry Pi
The main method for trying to connect to the Raspberry Pi was through Raspberry Pi’s official instruction guide for installing an operating system. (https://learn.adafruit.com/raspberry-pi-zero-creation/install-os-on-to-sd-card).
The instructions were clearly stated, and explicitly followed for each test—SD drive was erased and reformatted and the Pi was power cycled between every test. Raspbian Lite, Raspbian Stretch and NOOBS were as OS for the Raspberry Pi—none were able to help the Pi connect to a user interface.
Many methods were used to try and connect to the Raspberry Pi Zero W. SD card was formatted using various methods and then retested – SD card formatter was used, the mac disk utility was used to switch between EX-FAT, MS-DOS (FAT), MAC OS Journal, however, the partitioning appeared not to be the issue. Once disks were properly formatted, the chosen OS was burned to the SD card using the .img file and etcher. Each written image would be completed with a success. (The SD was determined not to be faulty as it was able to store and transport files properly).
After the image was properly written, the text files wpa_supplication.conf, config.txt, and ssh were either edited or created (where applicable). For Wi-Fi, the proper SSID and passwords were utilized for each network. After all the file editing was complete, the SD card was ejected properly, and placed into the micro-SD slot of the Raspberry Pi.
We tried connecting to the device through a mutually-shared network (both Pi and computer connected to the same network), however, the device would not connect to Wi-Fi, neither the BU (802.1x) when configured properly with IT, nor an iPhone hotspot, nor Steven Maloney’s PC hub hotspot.
Multiple USB-A to Micro-USB cords were utilized to connect to the Raspberry Pi using both the “USB” port and the “POWER IN” port as well. When the SD was plugged in, the Pi would indicate that it was reading the SD card  by flashing the green light many times before becoming constant. In no test were we able to connect and ping to the device—the error message would always read something along the lines of “could not locate raspberrypi.local”.
The Pi was tested using 4 different computers (2 PC, 2 Macs) for SD setup and connection. TA Steven Maloney, Jon-Michael Washington (a very intelligent CE), were contacted and were unable to connect to the Pi. After numerous hours spent trying to troubleshoot the Raspberry Pi, it was determined on Sunday night that the Pi was simply non-functional. 
The code we would’ve used would’ve been from https://raspberrypi4u.blogspot.com/2018/06/raspberry-pi-car-detection-with-opencv.html. It’s an open source and the youtube video displaying results were fairly accurate at detecting cars. We were unable to test and optimize the code due to our inability to test the program on a functional Raspberry Pi.
Reasons for not choosing the Software Project
Once we realized the Raspberry Pi would not work and that there were no other resources left, we decided to attempt to pursuit the software project. However, this was realized on Sunday, and it was determined, as electrical engineering, that our coding experience was not enough to complete such a task in 3 days, especially since we do not know how to develop websites and apps. Due to the uncertainty regarding an extension, this idea was dropped.
Individual Work Distribution
Cassie completed the software project code as well as some debug work.
Andrew completed the work and completed most of the Pi debugging work